Updating test code to remove "Test" from classes that was causing pytest to fail at command line.
Removing flake8 version from requirements_test.txt to make it compatible with python 3.7.x testing.  This is OK since automation will run flake8, and it's installed here only for local testing when automation fails due to linting.

Tweaking github actions workflow for python 3.7 for now to get that to pass.